Rename generateUtilities to substituteTailwindUtilitiesAtRules
